Added basic pinax layout.

+++ fabfile.py
@@ -11,9 +11,13 @@ def create_env():
     with cd(env.project_dir):
         if not dir_exists('env'):
             run('virtualenv env')
-        run('env/bin/pip install -r requirements.txt')
 
-        if not file_exists('local_settings.py'):
+        fabfile = 'requirements/{0}.txt'.format(env.type)
+        if file_exists(fabfile):
+            run('env/bin/pip install -r ' + fabfile)
+
+        if not file_exists('local_settings.py') \
+                and file_exists('configs/local_settings/{0}.py'.format(env.type)):
             run('ln -s configs/local_settings/{0}.py local_settings.py'.format(env.type))
 
 def _manage(command):
@@ -25,3 +29,4 @@ def _manage(command):
 shell = _manage('shell')
 dbshell = _manage('dbshell')
 runserver = _manage('runserver 0.0.0.0:8000')
+syncdb = _manage('syncdb')
